Refactor - change price type to Double

+++ src/main/java/com/digitalstork/devbookskata/service/DevelopmentBookService.java
@@ -9,5 +9,5 @@ public interface DevelopmentBookService {
 
     List<DevelopmentBookListDto> getAllDevelopmentBooks();
 
-    Integer purchaseBooks(List<DevelopmentBookPurchaseDto> purchaseDtos);
+    Double purchaseBooks(List<DevelopmentBookPurchaseDto> purchaseDtos);
 }
